This is a Black Scholes pricing model that shows how call and put prices vary when the stock prices and volatility are changes, this is represented on a heat map.
This project implements the Black-Scholes model for pricing European call and put options, featuring a Dash-based web app used for visualizing option prices. It calculates option values using the Black-Scholes formula, with the help of numpy and scipy.stats for efficient computations. The dashboard, built with Dash and Plotly, allows users to modify key parameters such as stock price, volatility, time to maturity, and risk-free rate, dynamically updating the visualizations. The interactive interface allows for a clear understanding of how option prices respond to market conditions.